backend/pyproject.toml
[tool.poetry]
name = "fastapi-netcom"
version = "0.1.0"
description = "FastAPI Network Commander"
authors = ["Ryan Bradshaw <ryan@rbradshaw.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.55.1"
uvicorn = "^0.11.5"
gunicorn = "^20.0.4"
asyncpg = "^0.20.1"
tortoise-orm = "^0.16.12"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
pyjwt = "^1.7.1"
pydantic = {extras = ["email"], version = "^1.5.1"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
isort = "^4.3.21"
pytest = "^5.4.3"
pytest-cov = "^2.10.0"
requests = "^2.23.0"
asynctest = "^0.13.0"

[tool.coverage.run]
omit = ["tests/*"]
branch = true

[tool.black]
line-length = 120
target-version = ['py38']

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

